Initial commit
This is still very much in a proof of concept state, though
the project should be in a good enough state that future
additions are relatively painless.

Text formatting nodes are not using the proper form in the
AST at the moment. They should not be relied upon just yet.